irc::connection: Abandon split send/recv design
When I first wrote the `irc::connection` module, the sole function of
a connection object was to split into a sender object and a receiver
object, patterned after the `SyncSender` and `Receiver` of
`std::sync::mpsc`.

I soon abandoned the idea of patterning the whole `irc` module after
`mpsc`, but the sender/receiver split lingered on, making the
higher-level interfaces a bit more cumbersome than necessary (at least
to implement).

This patch abandons the split sender/receiver design in favor of
having the connection object itself provide the `recv` and `try_send`
methods.