irc: Don't use usize::MAX as a Mio token value
Don't use `std::usize::MAX` as a Mio `Token` value to indicate that a
Mio event is associated with the MPSC queue via which the library
consumer can asynchronously send messages and other actions to this
library.

The reason for this change is that Mio itself uses `std::usize::MAX`
as a special token value, and errors out if its consumer (such as this
library) tries to do so. (How did this ever work? Mio seems to have
behaved thus for a while.)

Instead, define an enum to represent the possible sources of Mio
events (i.e., the MPSC queue and IRC sessions), and map between this
enum and Mio token values, assigning the MPSC queue token value `0`
rather than `std::usize::MAX`.

This mapping could certainly be performed without the enum, but I
think it adds a bit of safety.